requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
twilio>=9.0.0
tenacity>=8.2.0
python-multipart>=0.0.9
pytest>=8.0.0
httpx>=0.27.0
